+++ Common/src/main/java/com/tristankechlo/improvedvanilla/commands/ImprovedVanillaCommand.java
@@ -5,7 +5,9 @@ import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
 import com.tristankechlo.improvedvanilla.Constants;
 import com.tristankechlo.improvedvanilla.config.util.ConfigManager;
+import net.minecraft.ChatFormatting;
 import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.network.chat.Component;
 
 import static net.minecraft.commands.Commands.literal;
 
@@ -16,7 +18,13 @@ public class ImprovedVanillaCommand {
                 .then(literal("config").requires((source) -> source.hasPermission(3))
                         .then(literal("reload").executes(ImprovedVanillaCommand::configReload))
                         .then(literal("show").executes(ImprovedVanillaCommand::configShow))
-                        .then(literal("reset").executes(ImprovedVanillaCommand::configReset)));
+                        .then(literal("reset").executes(ImprovedVanillaCommand::configReset)))
+                .then(literal("github").executes(ImprovedVanillaCommand::github))
+                .then(literal("issue").executes(ImprovedVanillaCommand::issue))
+                .then(literal("wiki").executes(ImprovedVanillaCommand::wiki))
+                .then(literal("discord").executes(ImprovedVanillaCommand::discord))
+                .then(literal("curseforge").executes(ImprovedVanillaCommand::curseforge))
+                .then(literal("modrinth").executes(ImprovedVanillaCommand::modrinth));
         dispatcher.register(command);
         Constants.LOGGER.info("Command '/{}' registered", Constants.MOD_ID);
     }
@@ -41,5 +49,52 @@ public class ImprovedVanillaCommand {
         return 1;
     }
 
+    private static int github(CommandContext<CommandSourceStack> context) {
+        CommandSourceStack source = context.getSource();
+        Component link = ResponseHelper.clickableLink(Constants.GITHUB_URL);
+        Component message = Component.literal("Check out the source code on GitHub: ").withStyle(ChatFormatting.WHITE).append(link);
+        ResponseHelper.sendMessage(source, message, false);
+        return 1;
+    }
+
+    private static int issue(CommandContext<CommandSourceStack> context) {
+        CommandSourceStack source = context.getSource();
+        Component link = ResponseHelper.clickableLink(Constants.GITHUB_ISSUE_URL);
+        Component message = Component.literal("If you found an issue, submit it here: ").withStyle(ChatFormatting.WHITE).append(link);
+        ResponseHelper.sendMessage(source, message, false);
+        return 1;
+    }
+
+    private static int wiki(CommandContext<CommandSourceStack> context) {
+        CommandSourceStack source = context.getSource();
+        Component link = ResponseHelper.clickableLink(Constants.GITHUB_WIKI_URL);
+        Component message = Component.literal("The wiki can be found here: ").withStyle(ChatFormatting.WHITE).append(link);
+        ResponseHelper.sendMessage(source, message, false);
+        return 1;
+    }
+
+    private static int discord(CommandContext<CommandSourceStack> context) {
+        CommandSourceStack source = context.getSource();
+        Component link = ResponseHelper.clickableLink(Constants.DISCORD_URL);
+        Component message = Component.literal("Join the Discord here: ").withStyle(ChatFormatting.WHITE).append(link);
+        ResponseHelper.sendMessage(source, message, false);
+        return 1;
+    }
+
+    private static int curseforge(CommandContext<CommandSourceStack> context) {
+        CommandSourceStack source = context.getSource();
+        Component link = ResponseHelper.clickableLink(Constants.CURSEFORGE_URL);
+        Component message = Component.literal("Check out the CurseForge page here: ").withStyle(ChatFormatting.WHITE).append(link);
+        ResponseHelper.sendMessage(source, message, false);
+        return 1;
+    }
+
+    private static int modrinth(CommandContext<CommandSourceStack> context) {
+        CommandSourceStack source = context.getSource();
+        Component link = ResponseHelper.clickableLink(Constants.MODRINTH_URL);
+        Component message = Component.literal("Check out the Modrinth page here: ").withStyle(ChatFormatting.WHITE).append(link);
+        ResponseHelper.sendMessage(source, message, false);
+        return 1;
+    }
 
 }